Add Ordinal Theory
As Ordinal Theory forms the basis of this entire ecosystem it seems reasonable to consider it's inclusion as OIP 1. While this index would be traditionally (based on other existing Improvement Proposal mechanisms) be used to define the IP, I strongly believe that an acknowledgement of the work that created the need for this is not only reasonable, but necessary.